requirements.txt
# Core framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Configuration
pydantic==2.10.3
pydantic-settings==2.6.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1       # required by FastAPI TestClient

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Dev utilities
python-dotenv==1.0.1
